requirements.txt
Django==5.2
djangorestframework==3.16.0
django-cors-headers==4.7.0
pandas==2.2.1
plotly-express==0.4.1
Pillow==10.2.0
tensorflow==2.15.0
python-dotenv==1.0.1
requests==2.31.0
torch==2.2.2
torchvision==0.17.2